Add external openlibm. This provides a libm for reenix.
Rust needs to have a bunch of symbols from libm to link. :-/

Openlibm is taken from the Julia project (http://julialang.org/) and
the source was downloaded from https://github.com/JuliaLang/openlibm
on 09/22/14. Modifications to the openlibm source files are limited to
the makefiles.

Copyright noticies for openlibm are in kernel/external/openlibm/LICENSE.md