Added a crappy btree to mm.
This is going to be used to let us search for the right slab allocator
to use to fulfill memory allocations. It is not very good, (no
balancing, no removal, failures are fatal) but it gets the job done and
will be used nowhere else. I needed to make this instead of using a
stdlib Map implementation because I don't have allocation working yet at
this point so I need to use one that can do the allocation explicitly.

I am not sure if this works yet. I first need to get the rest of the
system up to check.