Change stage 3 to point %gs at top of stack
Rust uses %gs:0x30 to store where the stack ends (this is due to that
being the offset in TLS struct under linux) and we need to put a value
there so we will run until we can set up the stack and actually deal
with this. Currently we set 0 and then start the stack 0x34 lower than
what we normally would.